na/app/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
itsdangerous
python-dotenv
requests
markdown
sse-starlette
